remove a bug appearing with stripped binaries
When ELFFile was instantiated with a binary that had no .symtab section
it was throwing an exception when building the symbol interval tree. Now
the SIT is built lazily when needed (e.g. when get_symbol_by_offset is
called).
In addition, a new function ELFFile.has_symtab allows checking whether
there is a symbol table at all.