Fix a serious bug preventing BBs from being disassembled
This commit fixes the formula for calculating the windows size
while disassembling basic blocks. It assumed that when (start_addr,
end_addr) is given, the latter address will point to a byte "one after"
the basic block. Since it's not the case, the length of a BB has to be
incremented by one.